requirements.txt
###### Requirements without Version Specifiers ######`
argparse
feedparser
parse
python-dateutil
timezones
rocketchat_API

###### Requirements with Version Specifiers ######`

